+++ examples/unixfs-cat-readme.rs
@@ -7,7 +7,7 @@ use tokio::io::AsyncWriteExt;
 
 #[tokio::main]
 async fn main() -> anyhow::Result<()> {
-    let ipfs: Ipfs = UninitializedIpfs::new().start().await?;
+    let ipfs: Ipfs = UninitializedIpfs::new().with_default().start().await?;
     ipfs.default_bootstrap().await?;
 
     let mut stream = ipfs