+++ examples/unixfs_get.rs
@@ -23,7 +23,11 @@ async fn main() -> anyhow::Result<()> {
 
     tracing_subscriber::fmt::init();
 
-    let ipfs: Ipfs = UninitializedIpfs::new().enable_mdns().start().await?;
+    let ipfs: Ipfs = UninitializedIpfs::new()
+        .with_default()
+        .with_mdns()
+        .start()
+        .await?;
 
     if opt.default_bootstrap {
         ipfs.default_bootstrap().await?;