+++ packages/rust-ipns/examples/record.rs
@@ -6,16 +6,10 @@ use rust_ipns::Record;
 fn main() -> std::io::Result<()> {
     let keypair = Keypair::generate_ed25519();
 
-    let record = Record::new(
-        &keypair,
-        b"/path/cid",
-        Duration::seconds(60),
-        0,
-        0,
-    )?;
+    let record = Record::new(&keypair, b"/path/cid", Duration::seconds(60), 0, 0)?;
 
     let peer_id = keypair.public().to_peer_id();
     record.verify(peer_id)?;
-    
+
     Ok(())
 }