+++ src/repo/datastore/sled.rs
@@ -105,14 +105,14 @@ impl DataStore for SledDataStore {
 
     async fn iter(&self) -> futures::stream::BoxStream<'static, (Vec<u8>, Vec<u8>)> {
         let db = self.get_db().to_owned();
-        
+
         let stream = async_stream::stream! {
             let iter = db.iter();
             for (k, v) in iter.flatten() {
                 yield (k.to_vec(), v.to_vec());
             }
         };
-        
+
         stream.boxed()
     }