+++ tests/kademlia.rs
@@ -64,7 +64,7 @@ async fn spawn_bootstrapped_nodes<const N: usize>() -> (Vec<Node>, Option<Foreig
     // make sure that the nodes are not actively connected to each other
     // and that we are actually going to be testing the DHT here
     // for node in &nodes {
-        // assert!([1usize, 2].contains(&node.connected().await.unwrap().len()));
+    // assert!([1usize, 2].contains(&node.connected().await.unwrap().len()));
     // }
 
     (nodes, None)