+++ packages/libp2p-relay-manager/CHANGELOG.md
@@ -1,3 +1,11 @@
+# 0.1.2
+- chore: Emit event when selected peer is not an explicit relay
+
+# 0.1.1
+- fix: Emit event on dial failure [PR 105]
+
+[PR 105]: https://github.com/dariusc93/rust-ipfs/pull/105
+
 # 0.1.0
 - feat: Initial implementation of relay manager [PR 96]